fix(Cargo.toml): specify '*' version for ffmpeg->ffmpeg-sys
This restores the ability to `cargo package`, which is useful for
vendoring git dependencies with workspaces.

Although usually '*' version specifiers are frowned upon, it might make
sense for a path dependency, as it's impossible to fetch another version
of a crate via a path, and the crates in this repository should always
stay compatible with each other.